pyproject.toml
# Copyright (c) Nex-AGI. All rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "NexHTML"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mineru[core]>=2.6.4",
    "prompt-toolkit>=3.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "langfuse>=2.0.0",
    "qrcode",
    "pdf2image",
    "playwright==1.56.0",
]
